+++ R/zzz.R
@@ -8,7 +8,6 @@ minVersionCompatible <- "0.13.0"
 #' @name shared-add
 #'
 #' @param study An OmicAnalyzer study created with \code{\link{createStudy}}
-#' @param overwrite Overwrite existing value
 #'
 #' @keywords internal
 NULL